Fixed for results displaying more than once when specifying a formatter
The modified line was not taking into account a :cli formatting option like this: '--format=documentation' because the .include? test compares the entire line and not just the '--format' option specifier.

This fix greps for the format specifier at the beginning of the string and will disregard the '=' sign and anything that follows.