Fix a RSpec formatter issue
Commit c3b5c658ef28b4feb41136c2ae9a9893efda58fc introduced parsing of
the .rspec file if present with a fallback to a default formatter but
didn't handle the case where a .rspec file exists but without any
formatter option in it. In that case, no formatter was used meaning
that users got no feedback of the tests rsesults.

This commit fixes this issue by using a default formatter if there is
no .rspec file and no `--format` option given through the `:cli` option
or if there is a .rspec file but without any `--format` option in it.