Use a default formatter only if notification option is true
As Andrew Nikolaev pointed out: "If the notification option is false,
then there is no need to force the default formatter to add. The
default formatter is only useful in combination with
`-f Guard::RSpec::Formatter::Notification`. In all other cases it is
better to leave the usual behavior of rspec."