Refuse to run unless the cmd option is specified
After much discussion on the topic and many support questions about
bundler not being used it was decided that the default behavior should
be to not run any specs unless the user has specified exactly how they
should be run. In addition to warning the user of the change a sensible
default has been added to the Guardfile template.

Signed-off-by: Tyler Pickett <t.pickett66@gmail.com>